move moduleconfig to module package
