+++ src/main/java/uk/co/unitycoders/pircbotx/LocalConfiguration.java
@@ -20,6 +20,8 @@ package uk.co.unitycoders.pircbotx;
 
 import java.util.Map;
 
+import uk.co.unitycoders.pircbotx.modules.ModuleConfig;
+
 public class LocalConfiguration {
 
     public String nick;
