+++ src/main/java/uk/co/unitycoders/pircbotx/modules/ModuleConfig.java
@@ -1,4 +1,4 @@
-package uk.co.unitycoders.pircbotx;
+package uk.co.unitycoders.pircbotx.modules;
 
 public class ModuleConfig {
 	public String[] aliases;
